requirements.txt
# Python3.9+, update at 2023-07-01

# PyTorch
# numpy==1.25.1  # contain in torch
torch==2.0.1
torchvision==0.15.2

# d2l book，最新的0.17.6版本有冲突
# d2l==0.17.0

# for majsoul
mitmproxy==9.0.1
protobuf==3.16.0
requests
